+++ src/backend/protocols/go.smtp/cmd/caliopen_lmtpd/cli_cmds/root.go
@@ -16,7 +16,7 @@ var (
 	}
 )
 
-const __version__ = "0.13.2"
+const __version__ = "0.14.0"
 
 func init() {
 	cobra.OnInitialize()