+++ src/backend/tools/py.CLI/caliopen_cli/__init__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-__version__ = '0.13.2'
+__version__ = '0.14.0'
 
 try:
     import pkg_resources